Added Io_Thermistor to get raw GPIO voltages
wip
